+++ rvt/rvt/utils.py
@@ -1,5 +1,4 @@
 from typing import Iterator
-from urllib.parse import parse_qs, urlparse
 
 from requests import Response, Session
